PARCIAL!: HATEOAS sigue siendo un asco pero va queriendo con los ajustes que quedaron documentados en okdocs, también todavía hay muchos archivos y métodos basura del intento anterior, ni hablar de la mierda hardcodeada/fea pero eso lo voy a dejar para el final como limpieza.
* Seguí con revisar los links de paginacion que se generan del orto si te paras en una página que no existe
* Después ponete a hacer un atributo para marcar el SELF del controller para generar selfs de elementos individuales

+++ Sero.Doorman/Sero.Doorman.Tests/Tests/ResourcesController/Get_ByFilter.cs
@@ -21,7 +21,7 @@ namespace Sero.Doorman.Tests.Controllers.Resources
         [InlineData("rce_code_04", 1, 20, nameof(Resource.Code), Order.ASC)]
         [InlineData("resource_code_", 1, 20, nameof(Resource.Code), Order.ASC)]
         [InlineData("Resource_code_", 1, 20, nameof(Resource.Code), Order.ASC)]
-        public async Task Success(string textSearch, int page, int pageSize, string sortBy, string orderBy)
+        public async Task Success(string textSearch, ushort page, ushort pageSize, string sortBy, string orderBy)
         {
             // Arrange
             var sortBySelector = ReflectionUtils.GeneratePropertySelector<Resource>(sortBy);
@@ -59,7 +59,7 @@ namespace Sero.Doorman.Tests.Controllers.Resources
         [InlineData("gfdEdUDYzOHkhpFM7kGKTMkVX8", -1, -1, "carlitos", "carlitos")]    // code demasiado largo
         [InlineData("asd asd", -1, -1, "carlitos", "carlitos")]    // code con espacios
         [InlineData("asd}asd", -1, -1, "carlitos", "carlitos")]    // code con caracter inválido
-        public async Task InvalidFilter__Throws_ArgumentException(string textSearch, int page, int pageSize, string sortBy, string orderBy)
+        public async Task InvalidFilter__Throws_ArgumentException(string textSearch, ushort page, ushort pageSize, string sortBy, string orderBy)
         {
             await Assert.ThrowsAsync<ArgumentException>(async () =>
             {
